// frontend/src/api/generated/models/player-spsummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * GESTALOKA API
 * マルチプレイ・テキストMMO - ゲスタロカ API
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { SPSubscriptionType } from './spsubscription-type'

/**
 * プレイヤーSP概要スキーマ（軽量版）
 * @export
 * @interface PlayerSPSummary
 */
export interface PlayerSPSummary {
  /**
   *
   * @type {number}
   * @memberof PlayerSPSummary
   */
  current_sp: number
  /**
   *
   * @type {SPSubscriptionType}
   * @memberof PlayerSPSummary
   */
  active_subscription: SPSubscriptionType | null
  /**
   *
   * @type {string}
   * @memberof PlayerSPSummary
   */
  subscription_expires_at: string | null
}